Utils.py: Added tree_print method
The new tree_print method takes any object and prints its attributes as
a nice tree structure. In the future this method will be expanded to
print a nice full featured tree of the entire gnucash XML document, but
for now it is super useful for debugging what the actual objects look
like.