Transaction.py: Add object for transactions
Added Transaction.py which is a basic data structure for containing
information about each transaction. At this time there is no support for
the slots attribute of the transactions, mainly as I don't understand it
and can't easily make a datatype for it yet.

Also changed gnucash2ledger.py to iterate only over transactions now as
the splits are all part of a transaction anyway.

There is also a nice little TODO section to keep me on task.