Utils.py: Display nicely formatted trees
A couple of things have changed, notably the new conditional at the
start of tree_print which checks to see which indent level we're at. If
the indent level is not zero we should print out a nice line that traces
the relationship between the parent element that was printed on the line
above to the current element.

It also turns out that keeping the tree flowing downwards can be
accomplished easily by adding our full vertical box drawing character
before every tab.

Plus, fixed an off-by-on bug where the box drawing character indicating the end of
an elements attributes would never be printed.