Account.py: Add get_fullname and other changes
get_fullname recursively builds the full name of the account. Could
probably be speed up, but that comes after everything is working.

Also changed some of the way the XML representation of the account is
parsed into an Account object. This way we actually extract the currency
and add a list of slots that were found during the initial parsing.
Still not sure how exactly to deal with the slots.