Utils.py: Add support for changing line endings
output() was concieved as a wrapper around print(), but when lines using
print() were mass-converted to use output() the function call would fail
as a few calls to print() would use the print("text", end="") method to
print text without a newline.

This commit adds the ability to pass an ending to output() which will
default to the systems default line-ending, just like print() does.