Utils.py: Make sure element.text is not None
In some cases, it was possible to have a XML element that looked
something like this:

<trn:description></trn:descrption>

Which would cause safety_text to choke when trying to use replace() in
an attempt to sanitize the text elements read from GNUCash XML files.

This commit adds another check to guard against this condition and
return an empty string to that transactions may still be parsed.