Account.py: Remove "Root Account" from fullnames
get_fullname() used to include the "Root Account" as the beginning
of every fullname. This commit changes this.

Now, if the current account has no parent, then we know that this is the
root account and we can return None as its fullname so it is not
included in the recusive building of the fullname string.